rename components -> interfaces
